+++ tests/spawn.rs
@@ -31,7 +31,7 @@ fn empty() {
 
     let client_entity = client_app
         .world_mut()
-        .query_filtered::<Entity, With<Replicated>>()
+        .query_filtered::<Entity, With<Remote>>()
         .single(client_app.world())
         .unwrap();
 
@@ -70,7 +70,7 @@ fn with_component() {
     server_app.exchange_with_client(&mut client_app);
     client_app.update();
 
-    let mut components = client_app.world_mut().query::<(&Replicated, &A)>();
+    let mut components = client_app.world_mut().query::<(&Remote, &A)>();
     assert_eq!(components.iter(client_app.world()).count(), 1);
 }
 
@@ -99,7 +99,7 @@ fn with_multiple_components() {
     server_app.exchange_with_client(&mut client_app);
     client_app.update();
 
-    let mut components = client_app.world_mut().query::<(&Replicated, &A, &B)>();
+    let mut components = client_app.world_mut().query::<(&Remote, &A, &B)>();
     assert_eq!(components.iter(client_app.world()).count(), 1);
     assert_eq!(
         client_app.world().archetypes().len() - before_archetypes,
@@ -132,8 +132,8 @@ fn with_old_component() {
     client_app.update();
     server_app.exchange_with_client(&mut client_app);
 
-    let mut replicated = client_app.world_mut().query::<&Replicated>();
-    assert_eq!(replicated.iter(client_app.world()).len(), 0);
+    let mut remote = client_app.world_mut().query::<&Remote>();
+    assert_eq!(remote.iter(client_app.world()).len(), 0);
 
     // Enable replication for previously spawned entity
     server_app
@@ -145,7 +145,7 @@ fn with_old_component() {
     server_app.exchange_with_client(&mut client_app);
     client_app.update();
 
-    let mut components = client_app.world_mut().query::<(&Replicated, &A)>();
+    let mut components = client_app.world_mut().query::<(&Remote, &A)>();
     assert_eq!(components.iter(client_app.world()).count(), 1);
 }
 
@@ -173,8 +173,8 @@ fn empty_before_connection() {
     server_app.exchange_with_client(&mut client_app);
     client_app.update();
 
-    let mut components = client_app.world_mut().query::<&Replicated>();
-    assert_eq!(components.iter(client_app.world()).count(), 1);
+    let mut remote = client_app.world_mut().query::<&Remote>();
+    assert_eq!(remote.iter(client_app.world()).count(), 1);
 }
 
 #[test]
@@ -202,7 +202,7 @@ fn before_connection() {
     server_app.exchange_with_client(&mut client_app);
     client_app.update();
 
-    let mut components = client_app.world_mut().query::<(&Replicated, &A)>();
+    let mut components = client_app.world_mut().query::<(&Remote, &A)>();
     assert_eq!(components.iter(client_app.world()).count(), 1);
 }
 
@@ -246,7 +246,7 @@ fn signature() {
 
     let client_entity = client_app.world().entity(client_entity);
     assert!(
-        client_entity.contains::<Replicated>(),
+        client_entity.contains::<Remote>(),
         "entity should start receive replication"
     );
     assert!(
@@ -258,9 +258,9 @@ fn signature() {
         "component from server should be replicated"
     );
 
-    let mut replicated = client_app.world_mut().query::<&Replicated>();
+    let mut remote = client_app.world_mut().query::<&Remote>();
     assert_eq!(
-        replicated.iter(client_app.world()).count(),
+        remote.iter(client_app.world()).count(),
         1,
         "new entity shouldn't be spawned on client"
     );
@@ -486,8 +486,8 @@ fn hidden_entity() {
     client_app.update();
     server_app.exchange_with_client(&mut client_app);
 
-    let mut replicated = client_app.world_mut().query::<&Replicated>();
-    assert_eq!(replicated.iter(client_app.world()).len(), 0);
+    let mut remote = client_app.world_mut().query::<&Remote>();
+    assert_eq!(remote.iter(client_app.world()).len(), 0);
 }
 
 #[test]